Add pseudocode for building water molecules
Added pseudocode for building water molecules in the `build_h2o.pseudo` file. The pseudocode includes procedures for creating hydrogen and oxygen atoms, as well as a main procedure for coordinating the creation of atoms and forming water molecules. This code will be used as a starting point for implementing the functionality. Also added a `readme.adoc` file with instructions and explanations for the code.